bugfix change color inside usage of shadowdom
