+++ src/components/toolbars/toolbar/actions/color/color.tsx
@@ -3,6 +3,7 @@ import {Component, Event, EventEmitter, h, Prop, State} from '@stencil/core';
 import configStore from '../../../../../stores/config.store';
 import {ExecCommandAction} from '../../../../../types/execcommand';
 import {toHTMLElement} from '../../../../../utils/node.utils';
+import {getSelectionIncludingShadowroot} from '../../../../../utils/selection.utils.js';
 import {findStyleNode} from '../../../../../utils/toolbar.utils';
 
 @Component({
@@ -30,7 +31,7 @@ export class Color {
   }
 
   private initColor() {
-    const selection: Selection | null = getSelection();
+    const selection: Selection | null = getSelectionIncludingShadowroot(this.containerRef);
     this.range = selection?.getRangeAt(0);
 
     const anchor: HTMLElement | null = getAnchorElement(selection);
